Version 2.6
Feature: The support markers and the text displayed in the folds are now optional (unset by default), in a desperate attempt to achieve better performance on legacy systems.
If you want the opposite behavior then you have to set the globals g:ifold_support_markers and g:ifold_show_text to 1.